Use appropriate data type

+++ src/clause.py
@@ -85,5 +85,5 @@ class Clause:
 
         combined_columns = list(z_table.columns)[:-1] + list(pq_table.columns)[:-1]
         combined_array = np.array(combined_table)
-        combined_df = pd.DataFrame(combined_array, columns=combined_columns)
+        combined_df = pd.DataFrame(combined_array, columns=combined_columns).astype(np.uint8)
         return BitsTable(combined_df)
